@@ -37,46 +37,47 @@ def create_cohere_react_agent(
     prompt: ChatPromptTemplate,
 ) -> Runnable:
     """
-    Create an agent that enables multiple tools to be used in sequence to complete
-    a task.
+    Create an agent that enables multiple tools to be used in sequence to complete a
+    task.
 
     Args:
-        llm: The ChatCohere LLM instance to use.
+        llm: The `ChatCohere` LLM instance to use.
         tools: Tools this agent has access to.
         prompt: The prompt to use.
 
     Returns:
-        A Runnable sequence representing an agent. It takes as input all the same input
-        variables as the prompt passed in does and returns a List[AgentAction] or a
-        single AgentFinish.
+        A `Runnable` sequence representing an agent. It takes as input all the same input
+        variables as the prompt passed in does and returns a `list[AgentAction]` or a
+        single `AgentFinish`.
 
-        The AgentFinish will have two fields:
-        * output: str - The output string generated by the model
-        * citations: List[CohereCitation] - A list of citations that refer to the
+        The `AgentFinish` will have two fields:
+        * output: `str` - The output string generated by the model
+        * citations: `list[CohereCitation]` - A list of citations that refer to the
             output and observations made by the agent. If there are no citations this
             list will be empty.
 
     Example:
-        . code-block:: python
-            from langchain.agents import AgentExecutor
-            from langchain.prompts import ChatPromptTemplate
-            from langchain_cohere import ChatCohere, create_cohere_react_agent
+        ```python
+        from langchain.agents import AgentExecutor
+        from langchain.prompts import ChatPromptTemplate
+        from langchain_cohere import ChatCohere, create_cohere_react_agent
 
-            prompt = ChatPromptTemplate.from_template("{input}")
-            tools = []  # Populate this with a list of tools you would like to use.
+        prompt = ChatPromptTemplate.from_template("{input}")
+        tools = []  # Populate this with a list of tools you would like to use.
 
-            llm = ChatCohere()
+        llm = ChatCohere()
 
-            agent = create_cohere_react_agent(
-                llm,
-                tools,
-                prompt
-            )
-            agent_executor = AgentExecutor(agent=agent, tools=tools)
+        agent = create_cohere_react_agent(
+            llm,
+            tools,
+            prompt
+        )
+        agent_executor = AgentExecutor(agent=agent, tools=tools)
 
-            agent_executor.invoke({
-                "input": "In what year was the company that was founded as Sound of Music added to the S&P 500?",
-            })
+        agent_executor.invoke({
+            "input": "In what year was the company that was founded as Sound of Music added to the S&P 500?",
+        })
+        ```
     """  # noqa: E501
 
     # Creates a prompt, invokes the model, and produces a
@@ -102,10 +103,11 @@ def create_cohere_react_agent(
 
 
 class _AddCitations(Runnable):
-    """
-    Adds a list of citations to the output of the Cohere multi hop chain when the
-    last step is an AgentFinish. Citations are generated from the observations (made
-    in previous agent steps) and the grounded answer (made in the last step).
+    """Adds a list of citations to the output of the Cohere multi hop chain when the
+    last step is an `AgentFinish`.
+
+    Citations are generated from the observations (made in previous agent steps) and the
+    grounded answer (made in the last step).
     """
 
     def invoke(
