@@ -31,7 +31,7 @@ def enforce_stop_tokens(text: str, stop: List[str]) -> str:
 logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
-    from cohere.types import ListModelsResponse  # noqa: F401
+    from cohere import ListModelsResponse  # noqa: F401
 
 
 def completion_with_retry(llm: Cohere, **kwargs: Any) -> Any:
@@ -128,16 +128,16 @@ class BaseCohere(Serializable):
 class Cohere(LLM, BaseCohere):
     """Cohere large language models.
 
-    To use, you should have the ``cohere`` python package installed, and the
-    environment variable ``COHERE_API_KEY`` set with your API key, or pass
+    To use, you should have the `cohere` python package installed, and the
+    environment variable `COHERE_API_KEY` set with your API key, or pass
     it as a named parameter to the constructor.
 
     Example:
-        .. code-block:: python
+        ```python
+        from langchain_cohere import Cohere
 
-            from langchain_cohere import Cohere
-
-            cohere = Cohere(cohere_api_key="my-api-key")
+        cohere = Cohere(cohere_api_key="my-api-key")
+        ```
     """
 
     max_tokens: Optional[int] = None
@@ -150,14 +150,14 @@ class Cohere(LLM, BaseCohere):
     """Total probability mass of tokens to consider at each step."""
 
     frequency_penalty: Optional[float] = None
-    """Penalizes repeated tokens according to frequency. Between 0 and 1."""
+    """Penalizes repeated tokens according to frequency. Between `0` and `1`."""
 
     presence_penalty: Optional[float] = None
-    """Penalizes repeated tokens. Between 0 and 1."""
+    """Penalizes repeated tokens. Between `0` and `1`."""
 
     truncate: Optional[str] = None
     """Specify how the client handles inputs longer than the maximum token
-    length: Truncate from START, END or NONE"""
+    length: Truncate from `START`, `END` or `NONE`"""
 
     max_retries: int = 10
     """Maximum number of retries to make when generating."""
@@ -231,9 +231,9 @@ class Cohere(LLM, BaseCohere):
             The string generated by the model.
 
         Example:
-            .. code-block:: python
-
-                response = cohere("Tell me a joke.")
+            ```python
+            response = cohere("Tell me a joke.")
+            ```
         """
         params = self._invocation_params(stop, **kwargs)
         response = completion_with_retry(
@@ -259,9 +259,9 @@ class Cohere(LLM, BaseCohere):
             The string generated by the model.
 
         Example:
-            .. code-block:: python
-
-                response = await cohere("Tell me a joke.")
+            ```python
+            response = await cohere("Tell me a joke.")
+            ```
         """
         params = self._invocation_params(stop, **kwargs)
         response = await acompletion_with_retry(
