@@ -47,11 +47,11 @@ def _create_retry_decorator(max_retries: int) -> Callable[[Any], Any]:
 
 
 def _remove_signature_from_tool_description(name: str, description: str) -> str:
-    """
-    Removes the `{name}{signature} - ` prefix and Args: section from tool description.
-    The signature is usually present for tools created with the @tool decorator,
-    whereas the Args: section may be present in function doc blocks.
-    """
+    """Removes the `{name}{signature} - ` prefix and Args: section from tool description.
+
+    The signature is usually present for tools created with the `@tool` decorator,
+    whereas the `Args:` section may be present in function doc blocks.
+    """  # noqa: E501
     description = re.sub(rf"^{name}\(.*?\) -(?:> \w+? -)? ", "", description)
     description = re.sub(r"(?s)(?:\n?\n\s*?)?Args:.*$", "", description)
     return description
